move mock_network and rename it to mock_node
