+++ integration-tests/Cargo.toml
@@ -7,10 +7,6 @@ publish = false
 rust-version = "1.60.0"
 edition = "2021"
 
-[[bin]]
-path = "src/bin/start_mock_network.rs"
-name = "start_mock_network"
-
 [dependencies]
 actix = "0.13.0"
 actix-rt = "2"
@@ -25,8 +21,6 @@ once_cell = "1.5"
 parking_lot = "0.11.2"
 primitive-types = "0.10.1"
 rand = "0.7"
-rayon = "1.5"
-regex = "1"
 serde_json = "1"
 serde = { version = "1", features = ["derive"] }
 tempfile = "3"
@@ -99,4 +93,3 @@ sandbox = [
   "near-client/sandbox",
 ]
 no_cache = ["nearcore/no_cache"]
-mock_network = ["near-chain/mock_network", "near-epoch-manager/mock_network"]
